move timing functions to utility module
